+++ src/debug/DebugResolver.ts
@@ -1,30 +1,23 @@
-import {
-    DebugConfigurationProvider,
-    WorkspaceFolder,
-    CancellationToken,
-    DebugConfiguration,
-    ProviderResult,
-    debug
-} from 'vscode';
+import { debug, DebugConfiguration, DebugConfigurationProvider, ProviderResult, WorkspaceFolder } from 'vscode';
+import { getConfigFiles } from "../config/openocd";
 import { getConfig } from "../config/projectConfig";
 
 export default class DebugProvider implements DebugConfigurationProvider {
 
-    resolveDebugConfiguration?(folder: WorkspaceFolder | undefined, debugConfiguration: DebugConfiguration, token?: CancellationToken): ProviderResult<DebugConfiguration> {
+    resolveDebugConfiguration?(folder: WorkspaceFolder | undefined, debugConfiguration: DebugConfiguration): ProviderResult<DebugConfiguration> {
         console.log(debugConfiguration);
         if (debugConfiguration.type == 'kstm32') {
-            let config: DebugConfiguration = {
-                type: "cortex-debug",
-                name: "kstm32",
-                request: "launch",
-                cwd: "${workspaceRoot}",
-                executable: `./build/${getConfig().name || "invalid"}.elf`,
-                servertype: "openocd",
-                configFiles: [
-                    "kstm32-openocd-autogen.cfg"
-                ]
-            };
-            debug.startDebugging(folder, config);
+            getConfigFiles().then(configFiles => {
+                debug.startDebugging(folder, {
+                    type: "cortex-debug",
+                    name: "kstm32",
+                    request: "launch",
+                    cwd: "${workspaceRoot}",
+                    executable: `./build/${getConfig().name || "invalid"}.elf`,
+                    servertype: "openocd",
+                    configFiles: configFiles
+                } as DebugConfiguration);
+            });
         }
         return undefined;
     }