+++ CHANGELOG.md
@@ -6,6 +6,9 @@ All notable changes to the "kstm32" extension will be documented in this file
 ## [0.2.4]
 
 - 加入自动配置启动Cortex-Debug:OpenOCD的功能.
+- 移除运行OpenOCD的功能,改为调用Cortex-Debug插件.
+- 不再自动生成openocd配置文件
+- 不强制配置标准库路径
 
 
 ## [0.1.3]